group symbols, refactor Entra ID
