+++ packages/core/src/lib/actions/callback/oauth/callback.ts
@@ -17,7 +17,7 @@ import type {
 import { type OAuthConfigInternal } from "../../../../providers/index.js"
 import type { Cookie } from "../../../utils/cookie.js"
 import { isOIDCProvider } from "../../../utils/providers.js"
-import { fetchOpt } from "../../../utils/custom-fetch.js"
+import { conformInternal, customFetch } from "../../../symbols.js"
 import { decodeJwt } from "jose"
 
 function formUrlEncode(token: string) {
@@ -64,7 +64,7 @@ export async function handleOAuth(
     const issuer = new URL(provider.issuer!)
     const discoveryResponse = await o.discoveryRequest(issuer, {
       [o.allowInsecureRequests]: true,
-      ...fetchOpt(provider),
+      [o.customFetch]: provider[customFetch],
     })
     as = await o.processDiscoveryResponse(issuer, discoveryResponse)
 
@@ -167,7 +167,7 @@ export async function handleOAuth(
         if (!provider.checks.includes("pkce")) {
           args[1].body.delete("code_verifier")
         }
-        return fetchOpt(provider)[o.customFetch](...args)
+        return (provider[customFetch] ?? fetch)(...args)
       },
     }
   )
@@ -182,18 +182,33 @@ export async function handleOAuth(
 
   const requireIdToken = isOIDCProvider(provider)
 
-  if (provider.id === "microsoft-entra-id") {
-    const { tid } = decodeJwt((await codeGrantResponse.clone().json()).id_token)
-
-    if (typeof tid === "string") {
-      const tenantRe = /microsoftonline\.com\/(\w+)\/v2\.0/
-      const tenantId = as.issuer?.match(tenantRe)?.[1] ?? "common"
-      const issuer = new URL(as.issuer.replace(tenantId, tid))
-      const discoveryResponse = await o.discoveryRequest(
-        issuer,
-        fetchOpt(provider)
-      )
-      as = await o.processDiscoveryResponse(issuer, discoveryResponse)
+  if (provider[conformInternal]) {
+    switch (provider.id) {
+      case "microsoft-entra-id":
+      case "azure-ad": {
+        /**
+         * These providers need the authorization server metadata to be re-processed
+         * based on the `id_token`'s `tid` claim
+         * @see https://github.com/MicrosoftDocs/azure-docs/issues/113944
+         */
+        const { tid } = decodeJwt(
+          (await codeGrantResponse.clone().json()).id_token
+        )
+        if (typeof tid === "string") {
+          const tenantRe = /microsoftonline\.com\/(\w+)\/v2\.0/
+          const tenantId = as.issuer?.match(tenantRe)?.[1] ?? "common"
+          const issuer = new URL(as.issuer.replace(tenantId, tid))
+          const discoveryResponse = await o.discoveryRequest(issuer, {
+            [o.customFetch]: provider[customFetch],
+          })
+          as = await o.processDiscoveryResponse(issuer, discoveryResponse)
+        }
+        break
+      }
+      default:
+        throw new TypeError(
+          `Unrecognized provider conformation (${provider.id}).`
+        )
     }
   }
   const processedCodeResponse = await o.processAuthorizationCodeResponse(
@@ -218,7 +233,7 @@ export async function handleOAuth(
         client,
         processedCodeResponse.access_token,
         {
-          ...fetchOpt(provider),
+          [o.customFetch]: provider[customFetch],
           // TODO: move away from allowing insecure HTTP requests
           [o.allowInsecureRequests]: true,
         }
@@ -240,7 +255,7 @@ export async function handleOAuth(
         as,
         client,
         processedCodeResponse.access_token,
-        fetchOpt(provider)
+        { [o.customFetch]: provider[customFetch] }
       )
       profile = await userinfoResponse.json()
     } else {
