+++ packages/core/src/lib/actions/signin/authorization-url.ts
@@ -3,7 +3,7 @@ import * as o from "oauth4webapi"
 
 import type { InternalOptions, RequestInternal } from "../../../types.js"
 import type { Cookie } from "../../utils/cookie.js"
-import { fetchOpt } from "../../utils/custom-fetch.js"
+import { customFetch } from "../../symbols.js"
 
 /**
  * Generates an authorization/request token URL.
@@ -25,9 +25,9 @@ export async function getAuthorizationUrl(
     // We check this in assert.ts
 
     const issuer = new URL(provider.issuer!)
-    // TODO: move away from allowing insecure HTTP requests
     const discoveryResponse = await o.discoveryRequest(issuer, {
-      ...fetchOpt(options.provider),
+      [o.customFetch]: provider[customFetch],
+      // TODO: move away from allowing insecure HTTP requests
       [o.allowInsecureRequests]: true,
     })
     const as = await o.processDiscoveryResponse(issuer, discoveryResponse)
